herdr: Package for Animal Greenhouse Gas Emissions and Land Use

This package provides tools for calculating and analyzing greenhouse gas emissions and land use metrics for animal systems.

Details

For more information, see:

GitHub: https://github.com/JuanCBM99/herdr

See also

Author

Maintainer: Juan Carlos Baez juancarlos.baez@bc3research.org (ORCID)

Authors:

Jon Sampedro jon.sampedro@bc3research.org (ORCID)

Guillermo Pardo guillermo.pardo@bc3research.org (ORCID)

Agustín del Prado agustin.delprado@bc3research.org (ORCID)